车辆用户缴纳押金

+++ src/main/java/com/xiliulou/electricity/entity/EleDepositOrder.java
@@ -81,7 +81,7 @@ public class EleDepositOrder {
     /**
      * 门店Id
      */
-    private Integer storeId;
+    private Long storeId;
 
     /**
      * 车辆型号Id
